+++ R/check_zeroinflation.R
@@ -7,9 +7,13 @@
 #' @param x Fitted model of class `merMod`, `glmmTMB`, `glm`, or `glm.nb`
 #' (package **MASS**).
 #' @param tolerance The tolerance for the ratio of observed and predicted
-#'  zeros to considered as over- or underfitting zeros. A ratio
-#'  between 1 +/- `tolerance` is considered as OK, while a ratio
-#'  beyond or below this threshold would indicate over- or underfitting.
+#' zeros to considered as over- or underfitting zeros. A ratio
+#' between 1 +/- `tolerance` is considered as OK, while a ratio
+#' beyond or below this threshold would indicate over- or underfitting.
+#' @param alternative A character string specifying the alternative hypothesis.
+#' @param ... Arguments passed down to [`simulate_residuals()`]. This only applies
+#' for models with zero-inflation component, or for models of class `glmmTMB`
+#' from `nbinom1` or `nbinom2` family.
 #'
 #' @return A list with information about the amount of predicted and observed
 #'  zeros in the outcome, as well as the ratio between these two values.
@@ -19,14 +23,52 @@
 #' zero-inflation in the data. In such cases, it is recommended to use
 #' negative binomial or zero-inflated models.
 #'
+#' In case of negative binomial models, models with zero-inflation component,
+#' or hurdle models, the results from `check_zeroinflation()` are based on
+#' [`simulate_residuals()`], i.e. `check_zeroinflation(simulate_residuals(model))`
+#' is internally called if necessary.
+#'
+#' @section Tests based on simulated residuals:
+#' For certain models, resp. model from certain families, tests are based on
+#' [`simulated_residuals()`]. These are usually more accurate for tests than the
+#' traditionally used Pearson residuals. However, when simulating from more
+#' complex model, such as mixed models or models with zero-inflation, there are
+#' several important considerations. Arguments specified in `...` are passed to
+#' [`simulate_residuals()`], which relies on [`DHARMa::simulateResiduals()`] (and
+#' therefore, arguments in `...` are passed further down to _DHARMa_). The
+#' defaults in DHARMa are set on the most conservative option that works for
+#' all models. However, in many cases, the help advises to use different settings
+#' in particular situations or for particular models. It is recommended to read
+#' the 'Details' in `?DHARMa::simulateResiduals` closely to understand the
+#' implications of the simulation process and which arguments should be modified
+#' to get the most accurate results.
+#'
 #' @family functions to check model assumptions and and assess model quality
 #'
-#' @examplesIf require("glmmTMB")
+#' @examplesIf require("glmmTMB") && require("DHARMa")
 #' data(Salamanders, package = "glmmTMB")
 #' m <- glm(count ~ spp + mined, family = poisson, data = Salamanders)
 #' check_zeroinflation(m)
+#'
+#' # for models with zero-inflation component, it's better to carry out
+#' # the check for zero-inflation using simulated residuals
+#' m <- glmmTMB::glmmTMB(
+#'   count ~ spp + mined,
+#'   ziformula = ~ mined + spp,
+#'   family = poisson,
+#'   data = Salamanders
+#' )
+#' res <- simulate_residuals(m)
+#' check_zeroinflation(res)
 #' @export
-check_zeroinflation <- function(x, tolerance = 0.05) {
+check_zeroinflation <- function(x, ...) {
+  UseMethod("check_zeroinflation")
+}
+
+
+#' @rdname check_zeroinflation
+#' @export
+check_zeroinflation.default <- function(x, tolerance = 0.05, ...) {
   # check if we have poisson
   model_info <- insight::model_info(x)
   if (!model_info$is_count) {
@@ -41,28 +83,22 @@ check_zeroinflation <- function(x, tolerance = 0.05) {
     return(NULL)
   }
 
+  # model classes not supported in DHARMa
+  not_supported <- c("fixest", "glmx")
+
+  # for models with zero-inflation component or negative binomial families,
+  # we use simulated_residuals()
+  if (!inherits(x, not_supported) && (model_info$is_zero_inflated || model_info$is_negbin || model_info$family == "genpois")) { # nolint
+    if (missing(tolerance)) {
+      tolerance <- 0.1
+    }
+    return(check_zeroinflation(simulate_residuals(x, ...), tolerance = tolerance, ...))
+  }
+
   # get predictions of outcome
   mu <- stats::fitted(x)
-
-  # get overdispersion parameters
-  if (model_info$is_negbin) {
-    if (methods::is(x, "glmmTMB")) {
-      theta <- stats::sigma(x)
-    } else if (methods::is(x, "glmerMod")) {
-      theta <- environment(x@resp$family$aic)[[".Theta"]]
-    } else {
-      theta <- x$theta
-    }
-  } else {
-    theta <- NULL
-  }
-
   # get predicted zero-counts
-  if (!is.null(theta)) {
-    pred.zero <- round(sum(stats::dnbinom(x = 0, size = theta, mu = mu)))
-  } else {
-    pred.zero <- round(sum(stats::dpois(x = 0, lambda = mu)))
-  }
+  pred.zero <- round(sum(stats::dpois(x = 0, lambda = mu)))
 
   # proportion
   structure(
@@ -77,6 +113,33 @@ check_zeroinflation <- function(x, tolerance = 0.05) {
 }
 
 
+#' @rdname check_zeroinflation
+#' @export
+check_zeroinflation.performance_simres <- function(x,
+                                                   tolerance = 0.1,
+                                                   alternative = c("two.sided", "less", "greater"),
+                                                   ...) {
+  # match arguments
+  alternative <- match.arg(alternative)
+
+  # compute test results
+  result <- .simres_statistics(x, statistic_fun = function(i) sum(i == 0), alternative = alternative)
+
+  structure(
+    class = "check_zi",
+    list(
+      predicted.zeros = round(mean(result$simulated)),
+      observed.zeros = result$observed,
+      ratio = mean(result$simulated) / result$observed,
+      tolerance = tolerance,
+      p.value = result$p
+    )
+  )
+}
+
+#' @export
+check_zeroinflation.DHARMa <- check_zeroinflation.performance_simres
+
 
 # methods ------------------
 
@@ -90,12 +153,22 @@ print.check_zi <- function(x, ...) {
   lower <- 1 - x$tolerance
   upper <- 1 + x$tolerance
 
-  if (x$ratio < lower) {
-    message("Model is underfitting zeros (probable zero-inflation).")
-  } else if (x$ratio > upper) {
-    message("Model is overfitting zeros.")
+  if (is.null(x$p.value)) {
+    p_string <- ""
   } else {
-    insight::format_alert("Model seems ok, ratio of observed and predicted zeros is within the tolerance range.")
+    p_string <- paste0(" (", insight::format_p(x$p.value), ")")
+  }
+
+  if (x$ratio < lower) {
+    message("Model is underfitting zeros (probable zero-inflation)", p_string, ".")
+  } else if (x$ratio > upper) {
+    message("Model is overfitting zeros", p_string, ".")
+  } else {
+    insight::format_alert(paste0(
+      "Model seems ok, ratio of observed and predicted zeros is within the tolerance range",
+      p_string,
+      "."
+    ))
   }
 
   invisible(x)