+++ tests/testthat/test-check_autocorrelation.R
@@ -4,4 +4,9 @@ test_that("check_autocorrelation", {
   set.seed(123)
   out <- check_autocorrelation(m)
   expect_equal(as.vector(out), 0.316, ignore_attr = TRUE, tolerance = 1e-2)
+  expect_identical(
+    capture.output(print(out)),
+    "OK: Residuals appear to be independent and not autocorrelated (p = 0.316)."
+  )
+  expect_warning(plot(out), "There is currently")
 })