+++ tests/testthat/test-check_collinearity.R
@@ -202,14 +202,15 @@ test_that("check_collinearity, hurdle/zi models w/o zi-formula", {
     link = "logit"
   )
   out <- check_collinearity(m)
-  expect_identical(
-    colnames(out),
+  expect_named(
+    out,
     c(
       "Term", "VIF", "VIF_CI_low", "VIF_CI_high", "SE_factor", "Tolerance",
       "Tolerance_CI_low", "Tolerance_CI_high", "Component"
     )
   )
   expect_equal(out$VIF, c(1.05772, 1.05772, 1.06587, 1.06587), tolerance = 1e-4)
+  expect_snapshot(print(out))
 })
 
 test_that("check_collinearity, invalid data", {